Create env.py.orig
A frightening number Render build failures are due to the schema not being pre-pended after the student deletes the migration folder locally. I like the idea of this being saved in the outer directory structure so that it can be easily replaced in the migrations folder by a simple copy and paste operation in addition to being present and in their face as they work! Just a thought!! Thanks for indulging me!